+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 * Performance improvement: cached `git check-ignore` results.
 * Bug fix: correctly handle all Unicode characters.
 * Bug fix: when handling `workspace/didChangeWatchedFiles`, push the new diagnostics only once, after every file event has been processed.
-
+* Ignore binary files
 
 ## 0.0.1